refactor: move battle UI to helpers
